requirements-async.txt
# AI Company 非同期ワーカー専用依存関係
# 既存requirements.txtに追加する形で使用

# 非同期処理用
aio-pika>=9.0.0
aioredis>=2.0.0
aiofiles>=23.0.0

# 構造化ログ（軽量版）
structlog>=23.1.0

# メトリクス・監視
prometheus-client>=0.18.0

# サーキットブレーカー
circuitbreaker>=1.4.0

# セキュリティ・監視
cryptography>=41.0.0
watchdog>=3.0.0
psutil>=5.9.0

# 非同期テスト
pytest-asyncio>=0.21.0

# ユーティリティ
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0

# 既存パッケージ（確認用）
pika>=1.3.0
anthropic>=0.7.0
slack-sdk>=3.0.0
rich>=12.0.0
tabulate>=0.9.0
GitPython>=3.1.0